self-signed certificate support
- added checkbox to global config to allow plugin to ignore self-signed certificate errors, by default this is off
- removed legacy code which was ported into new more separated classes to allow easier testing
- added junit 5 dependencies to test scope
- updated LICENSE to match that proposed in original Rapid7 repository
- updated version in prep for 1.0.12 release
