+++ src/main/java/com/rapid7/appspider/Scan.java
@@ -17,12 +17,12 @@ public class Scan {
     private static final String FINISHED_SCANNING = SUCCESSFUL_SCAN + "|" + UNSUCCESSFUL_SCAN + "|" + FAILED_SCAN;
     private static final String UNAUTHORIZED_ERROR = "Unauthorized, please verify credentials and try again.";
 
-    private final EnterpriseClient client;
+    private final StandardEnterpriseClient client;
     private final ScanSettings settings;
     private final LoggerFacade log;
     private Optional<String> id;
 
-    public Scan(EnterpriseClient client, ScanSettings settings, LoggerFacade log) {
+    public Scan(StandardEnterpriseClient client, ScanSettings settings, LoggerFacade log) {
         if (Objects.isNull(client))
             throw new IllegalArgumentException("client cannot be null");
         if (Objects.isNull(settings))
@@ -75,12 +75,9 @@ public class Scan {
         }
 
         String status = client.getScanStatus(authToken.get(), runResult.getScanId()).orElse(FAILED_SCAN);
-        if (status.matches(SUCCESSFUL_SCAN)) {
-            log.println("Scan was complete but was not successful. Status was '" + status + "'");
-            return true;
-        }
-
-        log.println("Finished scanning!");
+        log.println(status.matches(SUCCESSFUL_SCAN)
+            ? "Scan was complete but was not successful. Status was '" + status + "'"
+            : "Finished scanning!");
         return true;
     }
